Only re-render files table values that have changed
Previously, each time we received a response from the /files API, we would re-render the entire table. We would also re-render the entire table whenever a user interacted with it. Now, we only re-render the cells that have changed since the previous render. This results in a performance improvement in general, and a massive improvement for torrents with many files.